Use format specifications in pack and unpack that are not affected by pointer size
`bytes_needed_buffer` is an integer of type DWORD, which is not affected by pointer size.
`module_handle_buffer` is an array of pointers. Using the 'J' specifier in the unpack format allows for reading based on the pointer size of the execution environment. The 'J' specifier has been available in Ruby since version 2.4.