Fixed to retrieve the error code from `$?`
In Ruby 1.8 and earlier, an integer representing the error code was assigned to `$?`. Starting from Ruby 1.9, it was changed to a `Process::Status` object. In the current Ruby version, the `case` statement in this part always executes the code under `else`. This was causing incorrect error messages to be displayed to users.