Added a comment about forcing encoding on `arg`
The code to force encoding on arg was implemented in commit a7e391d to address the issue mentioned in "Ocra fails when script is in a path with funky characters #42" (https://github.com/larsch/ocra/issues/42). This is a problem specific to Windows' Ruby 2.x as described in https://bugs.ruby-lang.org/issues/9016. While Ruby 3.x has resolved this issue, OCRAN still needs to support Ruby 2.6 and later.