Update gemspec required Ruby version to 2.6, include bundler with --add-all-core option, and fix test error
The test_rubyopt_manual test is failing due to the absence of the 'bundler' package in the test target. To include 'bundler' in the package for testing purposes, we will include the --add-all-core option in the package's options description.

Additionally, OCRAN officially supports Ruby 2.6 and higher versions. Since 'bundler' has become a default gem in Ruby 2.6 and above, we need to use the --add-all-core option. To reflect this change, we are updating the 'required_ruby_version' in the gemspec to 2.6.