Directly specify the path to the manifest file
The search for the manifest file involves a full directory tree traversal, which becomes time-consuming as the number of files increases. The required manifest file is specific to the RubyInstaller environment and its path can be determined. By directly specifying the path to the manifest file, unnecessary file searches are avoided.