Make OcranBuilder's stubimage write atomic
The block of code used for file writing has been changed to use `File.binwrite`. The writing of the stub image now involves closing the file once, which appears to be to avoid file access when appending icon resources to the stub file. This limitation is likely due to the use of the Win32API's `UpdateResource` function in `edicon.c`.
https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-updateresourcew