Added debug_extract and chdir_before to initialization options in OcranBuilder
OP_CREATE_INST_DIRECTORY must appear at the beginning of the data stream. Therefore, the create_inst_dir method has been made a private method, and it is called from the builder's initialization code. If this method is called multiple times, it will result in a failure of file extraction during stub execution. To avoid this, an exception is now raised if this method is called multiple times.