Create an ISS file in a temporary directory
When the output directory for the built installer file is not specified in the ISS file, Inno Setup uses the directory containing the ISS file as the working directory. To accommodate this, the location for creating temporary files has been set to the working directory at the time of command execution.