Eliminate duplicate target paths
Due to using the target argument's value as keys in a Hash, the intended
functionality of removing duplicates was not being performed. This is because
the Ocran::Pathname does not define a hash method. To address this, we now
generate keys for the hash on the C builder side, taking into account cases
where string paths are provided to the builder.